requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.46
databases==0.5.5
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.6.0
